backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "firebase-admin>=7.4.0",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.13",
    "langgraph>=1.1.6",
    "paddle-python>=0.6.0",
    "paddle-python-sdk>=1.14.0",
    "pydantic-settings>=2.13.1",
    "pygithub>=2.9.0",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
